% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hutcheson_T_test.R
\name{Hutcheson.T.test}
\alias{Hutcheson.T.test}
\title{Hutcheson T-test for two diversity indices.}
\usage{
Hutcheson.T.test(
  x,
  y,
  shannon.base = exp(1),
  alternative = "two.sided",
  difference = 0
)
}
\arguments{
\item{x, y}{Numeric vector of abundance of species for community \emph{x} and community \emph{y}}

\item{shannon.base}{A numeric indicating the logarithm base for the Shanon indices. Defaults to \emph{e}.}

\item{alternative}{A character indicating the alternative hypothesis. Can be "two.sided"(default), "less", "greater", or "auto"}

\item{difference}{A numeric indicating the value hypothesized of the difference between the indices. Defaults to 0.}
}
\value{
A list with class "htest" containing the following components:
\itemize{
\item statistic: Value of the Hutcheson T statistic
\item parameter: The degrees of freedom of the T statistic parameter
\item p.value: The test's p-value.
\item estimate: The Shannon diversity indices
\item null.value:The  hypothesized value of the difference between the Shannon diversty indices
\item method: Name of the test
\item alternative: The alternative hypothesis
\item data.name: Name of the data used in the test
}
}
\description{
This function performs Hutcheson (1970) to test the significance of the difference between two communities' diversity indices.
}
\details{
This function performs Hutcheson's t-test for comparing two diversity indices. This test is based on Shannon diversity indices' value computed using a logarithm base specified by the user. One-sided or two-sided tests are available.
}
\note{
missing values will be replaced with zero values
}
\examples{
# poner ejemplos, conseguir alguna base de datos que se publique con esto para usar aqui
# prueba una cola, etc
}
\references{
Zar, Jerrold H. 2010. Biostatistical Analysis. 5th ed. Pearson. pp. 174-176.

Hutcheson, Kermit. 1970. A Test for Comparing Diversities Based on the Shannon Formula. Journal of Theoretical Biology 29: 151-54.
}
\seealso{
See \code{\link[stats]{t.test}} in \pkg{stats} package.
}
\author{
David Ramirez Delgado \email{tucorreo@correo.com}.

Hugo Salinas \email{hugosal@comunidad.unam.mx}.
}
