% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_Hutcheson_t_test.R
\name{multiple_Hutcheson_t_test}
\alias{multiple_Hutcheson_t_test}
\title{Multiple Hutcheson t-tests comparisons between communities.}
\usage{
multiple_Hutcheson_t_test(x, shannon.base = exp(1))
}
\arguments{
\item{x}{Numeric dataframe or matrix of the abundance of species per
community. Each column corresponds to the communities and each row to
a species.}

\item{shannon.base}{A numeric indicating the logarithm base for computing
the Shannon indexes Defaults to \emph{e}.}
}
\value{
A matrix whose entries are the p-values of the test. The names
of the rows and columns are the names of the communities and its Shannon
diversity index.
}
\description{
This function computes the p-values of the Hutcheson (1970) to
test the significance of the difference between community diversity indexes.
}
\details{
This function performs Hutcheson's t-tests for comparing multiple
diversity indexes pairwise. This test is based on Shannon diversity indices'
value computed using a logarithm base specified by the user. The alternative
hypothesis is one-sided, chosen automatically according to the sign of the
difference between each pair of communities tested.
}
\note{
missing values will be replaced with zero values
}
\examples{
# poner algun ejemlpo aqui tambien
}
\seealso{
See \code{\link[ecolTest]{Hutcheson_t_test}} in \pkg{ecolTest}
package.
}
\author{
David Ramirez Delgado \email{tucorreo@correo.com}.

Hugo Salinas \email{hugosal@comunidad.unam.mx}.
}
