% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polychaeta_abundance}
\alias{polychaeta_abundance}
\title{Polychaeta abundance recorded in six samples.}
\format{
A data frame with 39 rows and 6 variables. Rownames are the name of
the species, the columns correspond to number of sample:
\describe{
\item{Sample.1}{The abundance of each species in the sample 1}
\item{Sample.2}{The abundance of each species in the sample 2}
\item{Sample.3}{The abundance of each species in the sample 3}
\item{Sample.4}{The abundance of each species in the sample 4}
\item{Sample.5}{The abundance of each species in the sample 5}
\item{Sample.6}{The abundance of each species in the sample 6}
}
}
\source{
xxxxxxxxxxxxxxxxxx articulo o publicacion
}
\usage{
polychaeta_abundance
}
\description{
A dataset containing the abundance of 38 species of Polychaeta worms
collected in XXXXXXXXXXXXXXXXX on XXXXXXXXXXXXXXXXXXX
by XXXXXXXXXXXXXXXXXXXXXXXXXXXXX
}
\keyword{datasets}
