% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{polychaeta_abundance}
\alias{polychaeta_abundance}
\title{Polychaeta abundance recorded in six samples.}
\format{
A data frame with 39 rows and 6 variables. Rownames are the name of
the species, the columns correspond to number of sample:
\describe{
\item{Sample.1}{The abundance of each species in the sample 1}
\item{Sample.2}{The abundance of each species in the sample 2}
\item{Sample.3}{The abundance of each species in the sample 3}
\item{Sample.4}{The abundance of each species in the sample 4}
\item{Sample.5}{The abundance of each species in the sample 5}
\item{Sample.6}{The abundance of each species in the sample 6}
}
}
\usage{
polychaeta_abundance
}
\description{
A dataset containing the abundance of 38 species of Polychaeta
collected in 2013 y 2014 from six samples from different locations in
Bahía de los Ángeles, Ensenada, Baja California, México, by
Victoria María Díaz Castañeda, PhD.  (Investigador Titular; Laboratorio de
Ecología del bentos marino, Departamentoe de Ecología Marina, CICESE).
}
\keyword{datasets}
